Shared columns now get the column type BLOB (instead of no type at all) to make obvious this directly uses the SQLite affinity BLOB (no mapping rule implied). Removed ECDbSqlColumn::Type::Long.

+++ iModelCore/ECDb/ECDb/ECDbMap.cpp
@@ -526,7 +526,7 @@ ECDbSqlTable* ECDbMap::FindOrCreateTable (SchemaImportContext* schemaImportConte
         if (Utf8String::IsNullOrEmpty (primaryKeyColumnName))
             primaryKeyColumnName = ECDB_COL_ECInstanceId;
 
-        auto column = table->CreateColumn (primaryKeyColumnName, ECDbSqlColumn::Type::Long, ColumnKind::ECInstanceId, PersistenceType::Persisted);
+        auto column = table->CreateColumn (primaryKeyColumnName, ECDbSqlColumn::Type::Integer, ColumnKind::ECInstanceId, PersistenceType::Persisted);
         if (table->GetPersistenceType () == PersistenceType::Persisted)
             {
             column->GetConstraintR ().SetIsNotNull (true);
@@ -535,9 +535,9 @@ ECDbSqlTable* ECDbMap::FindOrCreateTable (SchemaImportContext* schemaImportConte
 
         if (tableType == TableType::StructArray)
             {
-            table->CreateColumn (ECDB_COL_ParentECInstanceId, ECDbSqlColumn::Type::Long, ColumnKind::ParentECInstanceId, PersistenceType::Persisted);
-            table->CreateColumn (ECDB_COL_ECPropertyPathId, ECDbSqlColumn::Type::Long, ColumnKind::ECPropertyPathId, PersistenceType::Persisted);
-            table->CreateColumn (ECDB_COL_ECArrayIndex, ECDbSqlColumn::Type::Long, ColumnKind::ECArrayIndex, PersistenceType::Persisted);
+            table->CreateColumn (ECDB_COL_ParentECInstanceId, ECDbSqlColumn::Type::Integer, ColumnKind::ParentECInstanceId, PersistenceType::Persisted);
+            table->CreateColumn (ECDB_COL_ECPropertyPathId, ECDbSqlColumn::Type::Integer, ColumnKind::ECPropertyPathId, PersistenceType::Persisted);
+            table->CreateColumn (ECDB_COL_ECArrayIndex, ECDbSqlColumn::Type::Integer, ColumnKind::ECArrayIndex, PersistenceType::Persisted);
             if (table->GetPersistenceType() == PersistenceType::Persisted)
                 {
                 if (schemaImportContext != nullptr)
@@ -865,7 +865,7 @@ BentleyStatus ECDbMap::FinishTableDefinition () const
             if (addClassId)
                 {
                 const size_t insertPosition = 1;
-                ECDbSqlColumn * ecClassIdColumn = table->CreateColumn(ECDB_COL_ECClassId, ECDbSqlColumn::Type::Long, insertPosition, ColumnKind::ECClassId, PersistenceType::Persisted);
+                ECDbSqlColumn * ecClassIdColumn = table->CreateColumn(ECDB_COL_ECClassId, ECDbSqlColumn::Type::Integer, insertPosition, ColumnKind::ECClassId, PersistenceType::Persisted);
                 if (ecClassIdColumn == nullptr)
                     return ERROR;
